Check if osd is already registered before adding it to the auth list.
Conflicts:
	manifests/osd/device.pp
	spec/defines/ceph_osd_device_spec.rb

+++ manifests/osd/device.pp
@@ -106,6 +106,7 @@ size=1024m -n size=64k ${name}1",
         command => "\
 ceph auth add osd.${osd_id} osd 'allow *' mon 'allow rwx' \
 -i ${osd_data}/keyring",
+        unless  => "ceph auth list | egrep '^osd.${osd_id}$'",
         require => Exec["ceph-osd-mkfs-${osd_id}"],
       }
